fix magefiles
