+++ pkg/apis/kubernetes/client.go
@@ -1,6 +1,9 @@
 package client
 
 import (
+	"fmt"
+	"os"
+
 	toolchainv1alpha1 "github.com/codeready-toolchain/api/api/v1alpha1"
 	ecp "github.com/hacbs-contract/enterprise-contract-controller/api/v1alpha1"
 	ocpOauth "github.com/openshift/api/config/v1"
@@ -20,16 +23,23 @@ import (
 	"k8s.io/client-go/dynamic"
 	"k8s.io/client-go/kubernetes"
 	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
+	"k8s.io/client-go/rest"
+	"k8s.io/client-go/tools/clientcmd"
 	crclient "sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/config"
 )
 
-type K8sClient struct {
+type CustomClient struct {
 	kubeClient            *kubernetes.Clientset
 	crClient              crclient.Client
 	pipelineClient        pipelineclientset.Interface
 	dynamicClient         dynamic.Interface
 	jvmbuildserviceClient jvmbuildserviceclientset.Interface
+}
+
+type K8SClient struct {
+	AsKubeAdmin     *CustomClient
+	AsKubeDeveloper *CustomClient
 }
 
 var (
@@ -53,68 +63,93 @@ func init() {
 }
 
 // Kube returns the clientset for Kubernetes upstream.
-func (c *K8sClient) KubeInterface() kubernetes.Interface {
+func (c *CustomClient) KubeInterface() kubernetes.Interface {
 	return c.kubeClient
 }
 
 // Return a rest client to perform CRUD operations on Kubernetes objects
-func (c *K8sClient) KubeRest() crclient.Client {
+func (c *CustomClient) KubeRest() crclient.Client {
 	return c.crClient
 }
 
-func (c *K8sClient) PipelineClient() pipelineclientset.Interface {
+func (c *CustomClient) PipelineClient() pipelineclientset.Interface {
 	return c.pipelineClient
 }
 
-func (c *K8sClient) JvmbuildserviceClient() jvmbuildserviceclientset.Interface {
+func (c *CustomClient) JvmbuildserviceClient() jvmbuildserviceclientset.Interface {
 	return c.jvmbuildserviceClient
 }
 
 // Returns a DynamicClient interface.
 // Note: other client interfaces are likely preferred, except in rare cases.
-func (c *K8sClient) DynamicClient() dynamic.Interface {
+func (c *CustomClient) DynamicClient() dynamic.Interface {
 	return c.dynamicClient
 }
 
-// NewHASClient creates kubernetes client wrapper
-func NewK8SClient() (*K8sClient, error) {
-	cfg, err := config.GetConfig()
+func NewKubernetesClient() (*K8SClient, error) {
+	adminKubeconfig, err := config.GetConfig()
 	if err != nil {
 		return nil, err
 	}
 
-	client, err := kubernetes.NewForConfig(cfg)
+	adminClusterClient, err := createCustomClient(*adminKubeconfig)
 	if err != nil {
 		return nil, err
 	}
 
-	dynamicClient, err := dynamic.NewForConfig(cfg)
+	userKubeconfigPath := os.Getenv("USER_KUBE_CONFIG_PATH")
+	if userKubeconfigPath == "" {
+		return nil, fmt.Errorf("'USER_KUBE_CONFIG_PATH' env var needs to be exported and has to point to user kubeconfig generated by toolchain")
+	}
+
+	userCfg, err := clientcmd.BuildConfigFromFlags("", userKubeconfigPath)
 	if err != nil {
 		return nil, err
 	}
 
-	crClient, err := crclient.New(cfg, crclient.Options{
+	userClusterClient, err := createCustomClient(*userCfg)
+	if err != nil {
+		return nil, err
+	}
+
+	return &K8SClient{
+		AsKubeAdmin:     adminClusterClient,
+		AsKubeDeveloper: userClusterClient,
+	}, nil
+}
+
+func createCustomClient(cfg rest.Config) (*CustomClient, error) {
+	client, err := kubernetes.NewForConfig(&cfg)
+	if err != nil {
+		return nil, err
+	}
+
+	dynamicClient, err := dynamic.NewForConfig(&cfg)
+	if err != nil {
+		return nil, err
+	}
+
+	crClient, err := crclient.New(&cfg, crclient.Options{
 		Scheme: scheme,
 	})
 	if err != nil {
 		return nil, err
 	}
 
-	pipelineClient, err := pipelineclientset.NewForConfig(cfg)
+	pipelineClient, err := pipelineclientset.NewForConfig(&cfg)
 	if err != nil {
 		return nil, err
 	}
 
-	jvmbildserviceClient, err := jvmbuildserviceclientset.NewForConfig(cfg)
+	jvmbuildserviceClient, err := jvmbuildserviceclientset.NewForConfig(&cfg)
 	if err != nil {
 		return nil, err
 	}
-
-	return &K8sClient{
+	return &CustomClient{
 		kubeClient:            client,
 		crClient:              crClient,
 		pipelineClient:        pipelineClient,
-		jvmbuildserviceClient: jvmbildserviceClient,
 		dynamicClient:         dynamicClient,
+		jvmbuildserviceClient: jvmbuildserviceClient,
 	}, nil
 }
